rename `Preview` to `SafeBoundaries`
